Modify time management (seconds -> milliseconds)

+++ src/main/java/re/legend/crowd_simulator/graphics/MapListener.java
@@ -8,9 +8,9 @@ import re.legend.crowd_simulator.entities.bodies.AgentBody;
  *
  */
 public interface MapListener {
+	
 	/**
 	 * 
 	 */
-
-	void update(List<AgentBody> bodies, long time);
+	void update(List<AgentBody> bodies, float time);
 }
